+++ core/DataTable/Renderer/Json.php
@@ -73,7 +73,32 @@ class Json extends Renderer
         // silence "Warning: json_encode(): Invalid UTF-8 sequence in argument"
         $str = @json_encode($array);
 
+        if ($str === false
+            && json_last_error() === JSON_ERROR_UTF8
+            && $this->canMakeArrayUtf8()) {
+            $array = $this->makeArrayUtf8($array);
+            $str = json_encode($array);
+        }
+
         return $str;
+    }
+
+    private function canMakeArrayUtf8()
+    {
+        return function_exists('mb_convert_encoding');
+    }
+
+    private function makeArrayUtf8($array)
+    {
+        if (is_array($array)) {
+            foreach ($array as $key => $value) {
+                $array[$key] = self::makeArrayUtf8($value);
+            }
+        } elseif (is_string($array)) {
+            return mb_convert_encoding($array, 'UTF-8', 'auto');
+        }
+
+        return $array;
     }
 
     public static function sendHeaderJSON()