+++ core/Segment/SegmentExpression.php
@@ -301,9 +301,28 @@ class SegmentExpression
             }
         }
 
-        $this->checkFieldIsAvailable($field, $availableTables);
+        $columns = self::parseColumnsFromSqlExpr($field);
+        foreach ($columns as $column) {
+            $this->checkFieldIsAvailable($column, $availableTables);
+        }
 
         return array($sqlExpression, $value);
+    }
+
+    /**
+     * @param string $field
+     * @return string[]
+     */
+    public static function parseColumnsFromSqlExpr($field)
+    {
+        preg_match_all('/\b`?([a-zA-Z0-9_]+`?\.`?[a-zA-Z0-9_`]+)`?\b/', $field, $matches);
+        $result = isset($matches[1]) ? $matches[1] : [];
+        $result = array_map(function ($item) {
+            return str_replace('`', '', $item);
+        }, $result);
+        $result = array_unique($result);
+        $result = array_values($result);
+        return $result;
     }
 
     /**